requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
pandas
numpy
torch
torchvision
Pillow
python-dotenv
groq>=0.4.0
aiohttp
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
supabase
pydantic-settings
tenacity>=8.2.0
python-dateutil>=2.8.0
